refactor(game): move render into lib/render (T29)
Relocate render.ts and its colocated test into packages/game/lib/render/.
Imports of core/ modules become ../core/...; lib/loop.ts updated to
./render/render.js.
