+++ packages/game/lib/render/render.ts
@@ -1,4 +1,4 @@
-import type { GameState } from "./core/types.ts";
+import type { GameState } from "../core/types.ts";
 
 const render = (state: GameState, context: CanvasRenderingContext2D) => {
   // Draw player
