+++ specs/codebase-compliance/tasks.md
@@ -47,7 +47,7 @@ Each task ends in green `pnpm run ok` for the affected package(s)
 - [x] T26 — Move `packages/game/types/game-state-type.ts` → `packages/game/lib/core/types.ts`. Update every importer of the old path. `packages/game/types/env.d.ts` (the `__APP_VERSION__` ambient declaration) stays in `types/` — out of scope for layer rules. Used `git mv` to preserve history; nine importers updated to `./core/types.{ts,js}`. The `core/.gitkeep` placeholder is removed because `types.ts` now occupies the directory.
 - [x] T27 — Move `packages/game/lib/constants.ts` → `packages/game/lib/core/constants.ts`. Update importers. Seven source importers updated; `vitest.config.ts` coverage `exclude` list also updated to the new path so coverage thresholds remain enforced.
 - [x] T28 — Move `packages/game/lib/create-initial-state.ts`, `update-player.ts`, `update-enemies.ts`, `move-enemy-toward-player.ts` → `packages/game/lib/state/`. Update importers. Test files were moved alongside their sources (eight files total via `git mv`). Within-`state/` cross-imports stayed unchanged (still `./neighbour.js`); imports of `core/` modules became `../core/...`; the only outside importer was `lib/loop.ts`, updated to `./state/...`. `state/.gitkeep` removed.
-- [ ] T29 — Move `packages/game/lib/render.ts` → `packages/game/lib/render/render.ts`. Update importers.
+- [x] T29 — Move `packages/game/lib/render.ts` → `packages/game/lib/render/render.ts`. Update importers. `render.test.ts` moved alongside its source; the only outside importer was `lib/loop.ts`, updated to `./render/render.js`. `render/.gitkeep` removed.
 - [ ] T30 — Move `packages/game/lib/loop.ts`, `curtain-up.ts`, `bruff-game.ts`, `observable/` → `packages/game/lib/effects/` (preserving the `observable/` subdirectory). Update importers.
 - [ ] T31 — Add per-layer `no-restricted-imports` rule blocks to `@bruff/eslint-config` covering all five layers. Each block lists only the forbidden upward neighbours.
 - [ ] T32 — Run `pnpm run lint` from repo root; resolve any violation by relocating the offending file, not by widening the rule.
